:memo: refactor: rename recommediationRoutes to openaiRoutes
