+++ src/routes/openaiRoutes.js
